+++ documentation/Get-PnPAzureADActivityReportDirectoryAudit.md
@@ -20,7 +20,7 @@ Returns the audit logs generated by Azure AD.
 ## SYNTAX
 
 ```powershell
-Get-PnPAzureADActivityReportDirectoryAudit [-Identity <string>] [-Filter <string>]
+Get-PnPAzureADActivityReportDirectoryAudit [-Identity <string>] [-Filter <string>] [-Connection <PnPConnection>]
 ```
 
 ## DESCRIPTION
@@ -79,6 +79,21 @@ Specify the Filter condition for the audit log report.
 Type: string
 Parameter Sets: (All)
 Aliases:
+
+Required: False
+Position: Named
+Default value: None
+Accept pipeline input: False
+Accept wildcard characters: False
+```
+
+### -Connection
+
+Optional connection to be used by the cmdlet. Retrieve the value for this parameter by either specifying -ReturnConnection on Connect-PnPOnline or by executing Get-PnPConnection.
+
+```yaml
+Type: PnPConnection
+Parameter Sets: (All)
 
 Required: False
 Position: Named
