+++ src/Commands/Model/AzureAD/AzureADAuditAdditionalDetail.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace PnP.PowerShell.Commands.Model.AzureAD
+{
+    public class AzureADAuditAdditionalDetail
+    {
+        public string key { get; set; }
+        public string value { get; set; }
+    }
+}
