+++ src/Commands/Model/AzureAD/AzureADAuditInitiatedBy.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Text.Json.Serialization;
+
+namespace PnP.PowerShell.Commands.Model.AzureAD
+{
+    public class AzureADAuditInitiatedBy
+    {
+        [JsonPropertyName("user")]
+        public AzureADAuditUser User { get; set; }
+        [JsonPropertyName("app")]
+        public object app { get; set; }
+    }
+}
