+++ src/Commands/Model/AzureAD/AzureADAuditTargetResource.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Text.Json.Serialization;
+
+namespace PnP.PowerShell.Commands.Model.AzureAD
+{
+    public class AzureADAuditTargetResource
+    {
+        [JsonPropertyName("id")]
+        public string Id { get; set; }
+        [JsonPropertyName("displayName")]
+        public string DisplayName { get; set; }
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+        [JsonPropertyName("modifiedProperties")]
+        public List<AzureADAuditModifiedProperty> ModifiedProperties { get; set; }
+        [JsonPropertyName("groupType")]
+        public string GroupType { get; set; }
+        [JsonPropertyName("userPrincipalName")]
+        public string UserPrincipalName { get; set; }
+    }
+}
