+++ src/Commands/Model/AzureAD/AzureADAuditUser.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Text.Json.Serialization;
+
+namespace PnP.PowerShell.Commands.Model.AzureAD
+{
+    public class AzureADAuditUser
+    {
+        [JsonPropertyName("id")]
+        public string Id { get; set; }
+        [JsonPropertyName("displayName")]
+        public string DisplayName { get; set; }
+        [JsonPropertyName("userPrincipalName")]
+        public string UserPrincipalName { get; set; }
+        [JsonPropertyName("ipAddress")]
+        public string IPAddress { get; set; }
+    }
+}
